+++ zcash_client_sqlite/CHANGELOG.md
@@ -11,6 +11,7 @@ and this library adheres to Rust's notion of
 - `zcash_client_sqlite::WalletDb::with_gap_limits`
 - `zcash_client_sqlite::GapLimits`
 - `zcash_client_sqlite::util`
+- `zcash_client_sqlite::wallet::init::WalletMigrator`
 
 ### Changed
 - `zcash_client_sqlite::WalletDb` has an added `Clock` field and corresponding